requirements.txt
torch
transformers
pillow
numpy
matplotlib
opencv-python-headless
PyMuPDF
easyocr
timm
